v0.29.12.13

+++ Core.cpk/Scripts/Characters/Mobs/MobBossPragmiumQueen.cs
@@ -5,6 +5,7 @@
     using System.Linq;
     using AtomicTorch.CBND.CoreMod.Characters.Player;
     using AtomicTorch.CBND.CoreMod.CharacterSkeletons;
+    using AtomicTorch.CBND.CoreMod.Helpers.Server;
     using AtomicTorch.CBND.CoreMod.Items.Ammo;
     using AtomicTorch.CBND.CoreMod.Items.Weapons.MobWeapons;
     using AtomicTorch.CBND.CoreMod.SoundPresets;
@@ -64,8 +65,6 @@
 
         private const double VictoryLearningPointsBonusToEachAlivePlayer = 50;
 
-        private static readonly int MaxLootWinners;
-
         // Determines how often the boss will attempt to use the nova attack (a time interval in seconds).
         private static readonly Interval<double> NovaAttackInterval
             = new(min: 14, max: 17);
@@ -83,14 +82,25 @@
 
         private static readonly double ServerBossDifficultyCoef;
 
+        private static readonly int ServerMaxLootWinners;
+
         private IReadOnlyList<AiWeaponPreset> weaponsListNovaAttack;
 
         private IReadOnlyList<AiWeaponPreset> weaponsListPrimary;
 
         static MobBossPragmiumQueen()
         {
+            if (Api.IsClient)
+            {
+                return;
+            }
+
             var key = "BossDifficultyPragmiumQueen";
-            var defaultValue = 5.0; // by default the boss is balanced for 5 players
+            // by default the boss is balanced for 5 players
+            var defaultValue = Api.IsServer && ServerLocalModeHelper.IsLocalServer
+                                   ? 1.0
+                                   : 5.0;
+
             var description =
                 @"Difficulty of the Pragmium Queen (and the amount of loot/reward).
                   The number corresponds to the number of players necessary to kill the boss
@@ -117,8 +127,8 @@
             // coef range from 0.2 to 2.0
             ServerBossDifficultyCoef = requiredPlayersNumber / 5.0;
 
-            MaxLootWinners = (int)Math.Ceiling(requiredPlayersNumber * 2);
-            MaxLootWinners = Math.Max(MaxLootWinners, 5); // ensure at least 5 winners
+            ServerMaxLootWinners = (int)Math.Ceiling(requiredPlayersNumber * 2);
+            ServerMaxLootWinners = Math.Max(ServerMaxLootWinners, 5); // ensure at least 5 winners
         }
 
         public override bool AiIsRunAwayFromHeavyVehicles => false;
@@ -197,7 +207,7 @@
                                     lootObjectProto: ProtoLootObjectLazy.Value,
                                     lootObjectsDefaultCount: DeathSpawnLootObjectsDefaultCount,
                                     lootObjectsRadius: DeathSpawnLootObjectsRadius,
-                                    maxLootWinners: MaxLootWinners);
+                                    maxLootWinners: ServerMaxLootWinners);
                             }
                             finally
                             {
